+++ Classes/UVNewTicketViewController.h
@@ -12,27 +12,16 @@
 
 @class UVCustomField;
 
-@interface UVNewTicketViewController : UVBaseViewController 
-<UITableViewDataSource, UITableViewDelegate, 
-	UITextFieldDelegate, UVTextEditorDelegate> {
-		
-	NSString *text;
-	NSString *email;
-    NSString *subject;
+@interface UVNewTicketViewController : UVBaseViewController <UITableViewDataSource, UITableViewDelegate, UITextFieldDelegate, UVTextEditorDelegate> {
 	UVTextEditor *textEditor;
 	UITextField *emailField;
-    UITextField *subjectField;
 	UIBarButtonItem *prevBarButton;
     UIView *activeField;
 	//NSArray *customFields;
 }
 
-@property (nonatomic, retain) NSString *text;
-@property (nonatomic, retain) NSString *email;
-@property (nonatomic, retain) NSString *subject;
 @property (nonatomic, retain) UVTextEditor *textEditor;
 @property (nonatomic, retain) UITextField *emailField;
-@property (nonatomic, retain) UITextField *subjectField;
 @property (nonatomic, retain) UIBarButtonItem *prevBarButton;
 @property (nonatomic, retain) UIView *activeField;
 //@property (nonatomic, retain) NSArray *customFields;